% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning_functions.R
\docType{data}
\name{ARY}
\alias{ARY}
\alias{ary}
\alias{IRA}
\alias{ira}
\alias{ARARY}
\alias{arary}
\title{Learning curve functions and formulas}
\format{An object of class \code{formula} of length 3.}
\usage{
ARY

ary(ampl, rate, asym, trial)

IRA

ira(IRA, rate, ampl, trial)

ARARY

arary(amplS, rateS, amplM, rateM, asym, trialS, trialM)
}
\arguments{
\item{ampl}{amplitude (init - asym)}

\item{rate}{rate of learning}

\item{asym}{asymptote, max performance}

\item{trial}{practice trial}

\item{init}{initial performance}

\item{exp}{previous experience}
}
\value{
numeric

Computes the expected performance after practice trials using the exponential law of practice.
The following parametrizations are provided:

\describe{
\item{ary}{amplitude, rate, asymptote}
\item{ira}{initial performance, rate, amplitude}
}

In addition, regression formulas are provided for the Bayesian regression engine brms.
The right-hand side is always called \emph{perf}ormance.
By convention, formulas use upper case names, whereas the functions use lower-case.

Furthermore, two-component learning functions are provided for cases where performance is composed of
a general skill component \emph{S} (e.g. learning to draw a figure using a mirror)
and a motor sequence component \emph{M} (e.g., learning to draw the exact same figure)

\describe{
\item{arary}{amplitude_S, rate_S, amplitude_M, rate_M, asymptote}
}

All functions and formulas are also available with
link functions that put the parameters on a \emph{linear predictor scale, _lp} with a supported
range \eqn{[-\infty; \infty]}:

\describe{
\item{init, ampl, asym}{exp() <-> log()}
\item{rate}{inv_logit() <-> logit()}
}
}
\description{
exponential learning curve functions and non-linear regression formulas for use with brms
}
\examples{
ARY
ary(2, 0.3, 1, 1:5)

}
\author{
Martin Schmettow
}
\keyword{datasets}
