% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{coefary}
\alias{coefary}
\title{Quick estimation}
\usage{
smoothary(Laptrain$ToT, Laptrain$trial)
}
\arguments{
\item{perf}{}

\item{trial}{}
}
\value{
fitted values

Use nls() to estimate an ARY model and returns coefficients.
Trial counter and performance variable are given as vectors.
This makes smoothary useful in conjunction with manipulation chains.
With grouped mutation one can easily estimate a bunch of curves.
}
\description{
quickly estimates predicted values of an ARY model
}
\author{
Martin Schmettow
}
