% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{logit}
\alias{logit}
\alias{inv_logit}
\title{logit}
\usage{
logit(x)

inv_logit(x)
}
\arguments{
\item{x}{numerical}
}
\value{
numerical
}
\description{
computes the logit (or its inverse) on a numerical vector
}
\author{
Martin Schmettow
}
