% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{scor_to_ary}
\alias{scor_to_ary}
\title{Convert a SCOR model to ARY}
\usage{
scor_to_ary(posterior)
}
\arguments{
\item{posterior}{tbl_post object (bayr)}
}
\value{
tbl_post

The SCOR model often converges better, when estimated with brms.
Unfortunately, its offset parameter has rather little value. It is
converted to an asymptote parameter by offset * scale.
}
\description{
computes the asymptote parameter and adds it to the posterior
}
\author{
Martin Schmettow
}
