% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{set_response}
\alias{set_response}
\title{Set response variable}
\usage{
set_response(formula, response)
}
\arguments{
\item{formula}{learning curve formula}

\item{response}{response variable}
}
\value{
formula
}
\description{
sets the response variable of a formula
}
\author{
Martin Schmettow
}
