% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{smoothary}
\alias{smoothary}
\title{Smoother function}
\usage{
smoothary(Laptrain$ToT, Laptrain$trial)
}
\arguments{
\item{perf}{}

\item{trial}{}
}
\value{
fitted values

Use nls() to estimate an ARY model and returns fitted values.
With grouped mutation one can easily estimate a bunch of curves.
}
\description{
quickly estimates predicted values of an ARY model
}
\author{
Martin Schmettow
}
