% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning_functions.R
\docType{data}
\name{ARY}
\alias{ARY}
\alias{ary}
\alias{SCOR}
\alias{scor}
\alias{LSCOR}
\alias{lscor}
\alias{LARY}
\alias{lary}
\alias{iry}
\alias{ERY}
\alias{ery}
\alias{LERY}
\alias{lery}
\alias{ARARY}
\alias{arary}
\alias{LARARY}
\alias{larary}
\title{Learning curve functions and formulas}
\format{
An object of class \code{formula} of length 3.

An object of class \code{formula} of length 3.

An object of class \code{formula} of length 3.

An object of class \code{formula} of length 3.

An object of class \code{formula} of length 3.

An object of class \code{formula} of length 3.

An object of class \code{formula} of length 3.

An object of class \code{formula} of length 3.
}
\usage{
ARY

ary(ampl, rate, asym, trial)

SCOR

scor(scale, offset, rate, trial)

LSCOR

lscor(scale, offset, rate, trial)

LARY

lary(ampl, rate, trial)

iry(init, rate, asym, trial)

ERY

ery(pexp, rate, asym, trial)

LERY

lery(pexp, rate, asym, trial)

ARARY

arary(amplS, rateS, amplM, rateM, asym, trialS, trialM)

LARARY

larary(amplS, rateS, amplM, rateM, asym, trialS, trialM)
}
\arguments{
\item{ampl}{amplitude (init - asym)}

\item{rate}{rate of learning}

\item{asym}{asymptote, max performance}

\item{trial}{practice trial, starting at 0}

\item{init}{initial performance}

\item{exp}{previous experience}
}
\value{
numeric

Computes the expected performance after practice trials using the exponential law of practice.
The following parametrizations are provided:

\describe{
\item{ary}{ampl, rate, asym}
\item{ery}{pexp, rate, asym}
\item{iry}{init, rate, asym}
}

In addition, regression formulas are provided for the Bayesian regression engine brms.
The right-hand side is always called \emph{perf}ormance.
By convention, formulas use upper case names, whereas the functions use lower-case.

Furthermore, two-component learning functions are provided for cases where performance is composed of
a general skill component \emph{S} (e.g. learning to draw a figure using a mirror)
and a motor sequence component \emph{M} (e.g., learning to draw the exact same figure)

\describe{
\item{arary}{ampl, rate, asym}
}

All functions and formulas are also available with
link functions that put the parameters on a \emph{linear predictor scale} (supported
range \eqn{[-\infty; \infty]}) by log transformation. These formnulas/functions have the prefix \emph{L} (e.g., LARA)
}
\description{
exponential learning curve functions and non-linear regression formulas for use with brms
}
\examples{
ARY
ary(2, 0.3, 1, 1:5)
LARY
lary(log(2), log(0.3), log(1), 1:5)

}
\author{
Martin Schmettow
}
\keyword{datasets}
